requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.5
coloredlogs==15.0.1
dataclasses-json==0.5.9
decorator==5.1.1
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gitdb== 4.0.10
GitPython==3.1.32
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
langchain==0.0.235
langsmith==0.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pip==23.1.2
posthog==3.0.1
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
PyPDF2==3.0.1
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py== 0.8.11
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==67.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.19
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tesseract==0.1.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.3
wheel==0.38.4
yarl==1.9.2
zipp==3.16.2
